feat(zenko-739): Add Transient source test

+++ tests/create_buckets.py
@@ -42,7 +42,8 @@ buckets = [
     get_env('AZURE_BACKBEAT_SRC_CONTAINER_NAME', 'zenko-azure-crr-src-bucket'),
     get_env('WASABI_CRR_SRC_BUCKET', 'zenko-wasabi-crr-src-bucket'),
     get_env('DO_CRR_SRC_BUCKET', 'zenko-do-crr-src-bucket'),
-    get_env('MULTI_CRR_SRC_BUCKET', 'zenko-multi-crr-src-bucket')
+    get_env('MULTI_CRR_SRC_BUCKET', 'zenko-multi-crr-src-bucket'),
+    get_env('TRANSIENT_SRC_BUCKET', 'ci-zenko-transient-src-bucket')
 ]
 
 s = Session(aws_access_key_id=ZENKO_ACCESS_KEY,
@@ -84,16 +85,14 @@ while time.time() < timeout:
     ret = v1.list_namespaced_pod(K8S_NAMESPACE)
     passed = True
     for i in ret.items:
-        for container in i.status.container_statuses:
-            if not container.ready:
-                if container.state.waiting and \
-                    container.state.waiting.reason == 'CrashLoopBackOff':
-                    _log.info('%s is in CrashLoopBackOff, restarting.' % container.name)
-                    v1.delete_namespaced_pod(i.metadata.name, K8S_NAMESPACE, delete_opt)
-                passed = False
-                break
-        if not passed:
-            break
+        if i.status.container_statuses:
+            for container in i.status.container_statuses:
+                if not container.ready:
+                    if container.state.waiting and \
+                        container.state.waiting.reason == 'CrashLoopBackOff':
+                        _log.info('%s is in CrashLoopBackOff, restarting.' % container.name)
+                        v1.delete_namespaced_pod(i.metadata.name, K8S_NAMESPACE, delete_opt)
+                    passed = False
     if not passed:
         time.sleep(1)
     else:
